Move uploading project operators to new package
