+++ gridmarkets_blender_addon/__init__.py
@@ -47,7 +47,7 @@ modulesNames = ['meta_plugin.scene_exporter',
                 'property_groups.main_props',
                 'addon_preferences',
                 'blender_logging_wrapper',
-                'operators.add_remote_project',
+                'blender_plugin.remote_project_container.operators.add_remote_project',
                 'operators.trace_project',
                 'operators.submit',
                 'operators.submit_new_vray_project',
@@ -56,8 +56,8 @@ modulesNames = ['meta_plugin.scene_exporter',
                 'operators.open_help_url',
                 'operators.project_list_actions',
                 'operators.job_list_actions',
-                'operators.upload_project',
-                'operators.upload_file_as_project',
+                'blender_plugin.remote_project_container.operators.upload_project',
+                'blender_plugin.remote_project_container.operators.upload_file_as_project',
                 'operators.frame_range_list_actions',
                 'operators.vray_frame_range_list_actions',
                 'operators.edit_frame_range',
